requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
python-dotenv>=1.0.0
urllib3>=1.26.0

# Optional - for notifications
plyer>=2.1.0

# Optional - for PDF generation
reportlab>=4.0.0

# Optional - for FastAPI server
fastapi>=0.104.0
uvicorn>=0.24.0
openpyxl>=3.1.0

# For weekly report generation (DOCX & XLSX)
python-docx>=0.8.11